# nextjs-frontend/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "FastAPI",
		"version": "0.1.0"
	},
	"paths": {
...
		"/api/v1/items/": {
			"get": {
				"tags": ["items"],
				"summary": "Read Item",
				"operationId": "read_item",
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1,
							"description": "Page number",
							"default": 1,
							"title": "Page"
						},
						"description": "Page number"
					},
					{
						"name": "size",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"maximum": 100,
							"minimum": 1,
							"description": "Page size",
							"default": 50,
							"title": "Size"
						},
						"description": "Page size"
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Page_ItemRead_"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["items"],
				"summary": "Create Item",
				"operationId": "create_item",
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ItemCreate"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ItemRead"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/items/{item_id}": {
			"delete": {
				"tags": ["items"],
				"summary": "Delete Item",
				"operationId": "delete_item",
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"parameters": [
					{
						"name": "item_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Item Id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
			"ItemCreate": {
				"properties": {
					"name": {
						"type": "string",
						"title": "Name"
					},
					"description": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Description"
					},
					"quantity": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Quantity"
					}
				},
				"type": "object",
				"required": ["name"],
				"title": "ItemCreate"
			},
			"ItemRead": {
				"properties": {
					"name": {
						"type": "string",
						"title": "Name"
					},
					"description": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Description"
					},
					"quantity": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Quantity"
					},
					"id": {
						"type": "string",
						"format": "uuid",
						"title": "Id"
					},
					"user_id": {
						"type": "string",
						"format": "uuid",
						"title": "User Id"
					}
				},
				"type": "object",
				"required": ["name", "id", "user_id"],
				"title": "ItemRead"
			},
			"Page_ItemRead_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/ItemRead"
						},
						"type": "array",
						"title": "Items"
					},
					"total": {
						"anyOf": [
							{
								"type": "integer",
								"minimum": 0.0
							},
							{
								"type": "null"
							}
						],
						"title": "Total"
					},
					"page": {
						"anyOf": [
							{
								"type": "integer",
								"minimum": 1.0
							},
							{
								"type": "null"
							}
						],
						"title": "Page"
					},
					"size": {
						"anyOf": [
							{
								"type": "integer",
								"minimum": 1.0
							},
							{
								"type": "null"
							}
						],
						"title": "Size"
					},
					"pages": {
						"anyOf": [
							{
								"type": "integer",
								"minimum": 0.0
							},
							{
								"type": "null"
							}
						],
						"title": "Pages"
					}
				},
				"type": "object",
				"required": ["items", "page", "size"],
				"title": "Page[ItemRead]"
			},
...
			"ValidationError": {
				"properties": {
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"type": "array",
						"title": "Location"
					},
					"msg": {
						"type": "string",
						"title": "Message"
					},
					"type": {
						"type": "string",
						"title": "Error Type"
					}
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			},
...
		},
		"securitySchemes": {
			"OAuth2PasswordBearer": {
				"type": "oauth2",
				"flows": {
					"password": {
						"scopes": {},
						"tokenUrl": "auth/jwt/login"
					}
				}
			}
		}
	}
}